Only cp apidocs if api is generated.  Reference the previous comments on the trials with this PR.  John and I have been discussing the various alternatives and this approach looks to be the "final answer"...

+++ buildScripts/release.groovy
@@ -30,9 +30,14 @@ pipeline {
         }
         stage("Copy Specs") {
             steps {
-                sh "mkdir -p /home/data/httpd/download.eclipse.org/microprofile/${params.module}-${params.releaseVersion}/apidocs"
+                sh "mkdir -p /home/data/httpd/download.eclipse.org/microprofile/${params.module}-${params.releaseVersion}"
                 sh "cp -r spec/target/generated-docs/* /home/data/httpd/download.eclipse.org/microprofile/${params.module}-${params.releaseVersion}"
-                sh "cp -r api/target/apidocs/* /home/data/httpd/download.eclipse.org/microprofile/${params.module}-${params.releaseVersion}/apidocs"
+                script {
+                    if (fileExists('api')) {
+                        sh "mkdir -p /home/data/httpd/download.eclipse.org/microprofile/${params.module}-${params.releaseVersion}/apidocs"
+                        sh "cp -r api/target/apidocs/* /home/data/httpd/download.eclipse.org/microprofile/${params.module}-${params.releaseVersion}/apidocs"
+                    }
+                }
             }
         }
     }
